Remove signon_identity_verify_user()
The unit test was provoking a crash: the function signature is wrong,
because on signond side this method takes a dictionary parameter, while
the client API has a username parameters.
Resolve this conflict by removing this function; we'll add it back
later, with the correct signature, if and when needed.