AuthSession: bring back libsignon-glib -compatible constructor
Reintroduce the constructor which takes an integer parameter, to
increase compatibility with libsignon-glib, and rename the existing
constructor to signon_auth_session_new_for_identity().
Also, declare a couple of types in signon-types.h, so that the signature
of the signon_auth_session_new_for_identity() can now properly accept a
SignonIdentity * (same trick used in libaccounts-glib).